Seperated data structures & file handler
