+++ include/data.h
@@ -1,0 +1,31 @@
+/* 
+ * simple lennard-jones potential MD code with velocity verlet.
+ * units: Length=Angstrom, Mass=amu; Energy=kcal
+ *
+ * baseline c version.
+ */
+
+#ifndef DATA_H
+#define DATA_H
+
+#include <stdio.h>
+#include <ctype.h>
+#include <string.h>
+#include <stdlib.h>
+
+
+/* structure to hold the complete information 
+ * about the MD system */
+struct _mdsys {
+    int natoms,nfi,nsteps;
+    double dt, mass, epsilon, sigma, box, rcut;
+    double ekin, epot, temp;
+    double *rx, *ry, *rz;
+    double *vx, *vy, *vz;
+    double *fx, *fy, *fz;
+};
+typedef struct _mdsys mdsys_t;
+
+int get_a_line(FILE *fp, char *buf);
+ 
+#endif
