+++ src/data.c
@@ -1,0 +1,39 @@
+/* 
+ * simple lennard-jones potential MD code with velocity verlet.
+ * units: Length=Angstrom, Mass=amu; Energy=kcal
+ *
+ * baseline c version.
+ */
+
+#include "data.h"
+#define BLEN 200
+
+/* helper function: read a line and then return
+   the first string with whitespace stripped off */
+int get_a_line(FILE *fp, char *buf)
+{
+    char tmp[BLEN], *ptr;
+
+    /* read a line and cut of comments and blanks */
+    if (fgets(tmp,BLEN,fp)) {
+        int i;
+
+        ptr=strchr(tmp,'#');
+        if (ptr) *ptr= '\0';
+        i=strlen(tmp); --i;
+        while(isspace(tmp[i])) {
+            tmp[i]='\0';
+            --i;
+        }
+        ptr=tmp;
+        while(isspace(*ptr)) {++ptr;}
+        i=strlen(ptr);
+        strcpy(buf,tmp);
+        return 0;
+    } else {
+        perror("problem reading input");
+        return -1;
+    }
+    return 0;
+}
+
